homeassistant/addons/roamcore-victron-mock/requirements.txt
paho-mqtt>=2.1.0

